move logic to main.tsx and integrated into app.tsx
