+++ src/pages/Main/Main.module.css
@@ -1,4 +1,4 @@
-.app {
+.main {
   display: flex;
   flex-direction: column;
   height: 100vh;
